+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Lupa change log
 ===============
 
+Upcoming release
+----------------
+
+* The runtime version of the Lua library as a tuple (e.g. ``(5,3)``)
+  is provided via ``lupa.LUA_VERSION`` and ``LuaRuntime.lua_version``.
+
+
 1.9 (2019-12-21)
 ----------------